refactor: moved new functions to conform with existing naming conventions
see https://github.com/silverstripe/silverstripe-cms/pull/2703#discussion_r789208512
